+++ packages/eyes-testcafe/test/coverage/index.js
@@ -1,0 +1,6 @@
+const {coverageTestsConfig} = require('@applitools/sdk-shared')
+
+const modifiedConfig = {...coverageTestsConfig}
+modifiedConfig.template = './test/coverage/template.hbs'
+
+module.exports = modifiedConfig