+++ packages/sdk-shared/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added debugScreenshots boolean option to getEyes that will automatically create a screenshots dir in the cwd
 
 ## 0.7.1 - 2021/1/5
 
@@ -120,4 +121,3 @@
 - remove yarn workspaces
 - updated to @applitools/functional-commons@1.6.0 (from 1.5.4)
 - updated to @applitools/sdk-coverage-tests@1.0.14 (from 1.0.13)
-